+++ NEWS.md
@@ -12,6 +12,8 @@
 -   `cff_parse_citation()` does a better job extracting information from
     `bibentry()` /BibTeX and mapping it to `preferred-citation/references`
     fields of CFF.
+    
+-   New helper function `cff_parse_person_bibtex()`.
 
 # cffr 0.1.1